flatterers-dilemma/requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
nltk>=3.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wandb>=0.16.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
